move the parquet variant encode/decode into core
